Refactor JcloudsRebindWithYamlDslTest
Remove duplication from JcloudsRebindWithExternalConfigTest, creating
common abstract super-type.
