+++ camp/camp-brooklyn/src/test/java/org/apache/brooklyn/camp/brooklyn/JcloudsRebindWithExternalConfigTest.java
@@ -20,25 +20,20 @@ package org.apache.brooklyn.camp.brooklyn;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
-import java.io.File;
 import java.util.Map;
 
 import org.apache.brooklyn.api.entity.Application;
 import org.apache.brooklyn.api.entity.Entity;
 import org.apache.brooklyn.api.entity.EntitySpec;
-import org.apache.brooklyn.api.mgmt.ha.HighAvailabilityMode;
 import org.apache.brooklyn.camp.brooklyn.spi.creation.CampTypePlanTransformer;
 import org.apache.brooklyn.core.config.external.InPlaceExternalConfigSupplier;
 import org.apache.brooklyn.core.entity.trait.Startable;
 import org.apache.brooklyn.core.internal.BrooklynProperties;
-import org.apache.brooklyn.core.mgmt.internal.LocalManagementContext;
 import org.apache.brooklyn.core.typereg.RegisteredTypeLoadingContexts;
 import org.apache.brooklyn.location.jclouds.ComputeServiceRegistry;
 import org.apache.brooklyn.location.jclouds.JcloudsLocation;
 import org.apache.brooklyn.location.jclouds.JcloudsPropertiesFromBrooklynProperties;
-import org.apache.brooklyn.location.jclouds.JcloudsRebindStubTest;
 import org.apache.brooklyn.util.collections.MutableMap;
-import org.testng.annotations.AfterMethod;
 import org.testng.annotations.Test;
 
 import com.google.common.base.Joiner;
@@ -46,37 +41,14 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Iterables;
 
 /**
- * Implementation notes. This relies on the test {@link JcloudsRebindStubTest#testRebind()}.
  * It changes the setup for the test in the following ways:
  * <ul>
  *   <li>Brooklyn properties defines external config
- *   <li>Location is defined in YAML, and refers to the external config for the identity/credential.
- *   <li>When creating management context, it also creates {@link BrooklynCampPlatformLauncherNoServer}.
- *   <li>It uses {@link JcloudsRebindWithExternalConfigTest#ByonComputeServiceStaticRef} to allow
- *       the test's {@link ComputeServiceRegistry} to be wired up via YAML.
  * </ul>
- * 
- * See {@link JcloudsRebindStubTest} for explanation why this is "Live" - it will not create VMs,
- * but does retrieve list of images etc.
  */
 @Test(groups={"Live", "Live-sanity"})
-public class JcloudsRebindWithExternalConfigTest extends JcloudsRebindStubTest {
+public class JcloudsRebindWithExternalConfigTest extends JcloudsRebindStubYamlTest {
 
-    private BrooklynCampPlatformLauncherNoServer origLauncher;
-    private BrooklynCampPlatformLauncherNoServer newLauncher;
-
-    @Override
-    @AfterMethod(alwaysRun=true)
-    public void tearDown() throws Exception {
-        try {
-            super.tearDown();
-        } finally {
-            ByonComputeServiceStaticRef.clearInstance();
-            if (origLauncher != null) origLauncher.stopServers();
-            if (newLauncher != null) newLauncher.stopServers();
-        }
-    }
-    
     @Override
     protected BrooklynProperties createBrooklynProperties() {
         BrooklynProperties result = super.createBrooklynProperties();
@@ -93,39 +65,15 @@ public class JcloudsRebindWithExternalConfigTest extends JcloudsRebindStubTest {
     }
     
     @Override
-    protected LocalManagementContext createOrigManagementContext() {
-        origLauncher = new BrooklynCampPlatformLauncherNoServer() {
-            @Override
-            protected LocalManagementContext newMgmtContext() {
-                return JcloudsRebindWithExternalConfigTest.super.createOrigManagementContext();
-            }
-        };
-        origLauncher.launch();
-        LocalManagementContext mgmt = (LocalManagementContext) origLauncher.getBrooklynMgmt();
-        return mgmt;
-    }
-
-    @Override
-    protected LocalManagementContext createNewManagementContext(final File mementoDir, final HighAvailabilityMode haMode) {
-        newLauncher = new BrooklynCampPlatformLauncherNoServer() {
-            @Override
-            protected LocalManagementContext newMgmtContext() {
-                return JcloudsRebindWithExternalConfigTest.super.createNewManagementContext(mementoDir, haMode);
-            }
-        };
-        newLauncher.launch();
-        return (LocalManagementContext) newLauncher.getBrooklynMgmt();
-    }
-    
-    @Override
     protected JcloudsLocation newJcloudsLocation(ComputeServiceRegistry computeServiceRegistry) throws Exception {
         ByonComputeServiceStaticRef.setInstance(computeServiceRegistry);
         
         String yaml = Joiner.on("\n").join(
                 "location:",
-                "  jclouds:softlayer:",
+                "  "+LOCATION_SPEC+":",
                 "    identity: $brooklyn:external(\"creds\", \"test-identity\")",
                 "    credential: $brooklyn:external(\"creds\", \"test-credential\")",
+                "    imageId: "+IMAGE_ID,
                 "    jclouds.computeServiceRegistry:",
                 "      $brooklyn:object:",
                 "        type: "+ByonComputeServiceStaticRef.class.getName(),
@@ -140,19 +88,5 @@ public class JcloudsRebindWithExternalConfigTest extends JcloudsRebindStubTest {
         app.invoke(Startable.START, ImmutableMap.<String, Object>of()).get();
 
         return (JcloudsLocation) Iterables.getOnlyElement(app.getLocations());
-    }
-
-    public static class ByonComputeServiceStaticRef {
-        private static volatile ComputeServiceRegistry instance;
-
-        public ComputeServiceRegistry asComputeServiceRegistry() {
-            return checkNotNull(instance, "instance");
-        }
-        static void setInstance(ComputeServiceRegistry val) {
-            instance = val;
-        }
-        static void clearInstance() {
-            instance = null;
-        }
     }
 }
